pyproject.toml
[project]
name = "code-exec-plaftorm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.3",
    "docker>=7.1.0",
    "fastapi-users[sqlalchemy]>=15.0.5",
    "fastapi[standard]>=0.136.0",
    "pika>=1.3.2",
    "psycopg2>=2.9.11",
    "python-decouple>=3.8",
    "uvicorn>=0.44.0",
]
